Удалит genreName из тестовых моков, расширит моки для разработки.
