+++ src/components/genres-item/genres-item.e2e.test.js
@@ -1,13 +1,13 @@
 import React from "react";
 import {shallow} from "enzyme";
 import {GenresItem} from "./genres-item.jsx";
-import {genreName} from "../../__test-data__/test-mocks.js";
+import {genreNames} from "../../__test-data__/test-mocks.js";
 
 
 const onClick = jest.fn();
 
 const props = {
-  genreName,
+  genreName: genreNames[6],
   onClick,
 };
