+++ src/mocks/promo-movie.js
@@ -1,10 +1,7 @@
-const promoMovie = {
+export const promoMovie = {
   title: `The Grand Budapest Hotel`,
   genre: `Drama`,
   releaseDate: new Date(2014, 0),
   posterUrl: `img/the-grand-budapest-hotel-poster.jpg`,
   backgroundUrl: `img/bg-the-grand-budapest-hotel.jpg`,
 };
-
-
-export {promoMovie};
